Updated About section.
