--- client/src/pages/About/Login.css
@@ -1,11 +1,0 @@
-.loginTitle {
-	font-size: 5em;
-	letter-spacing: 10px;
-	line-height: 150%;
-	text-shadow:
-		3px 3px 0 #034,
-		-3px -3px 0 #034,  
-		3px -3px 0 #034,
-		-3px 3px 0 #034,
-		3px 3px 0 #034;
-}
