+++ client/src/pages/About/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Login.js";
+export { default } from "./About.js";
